[Ref/#131] QnaDocument 제거 및 QnaDocument1의 클래스명을 QnaDocument로 변경
