+++ app/src/main/java/com/abloom/mery/data/firebase/qna/QnaDocument.kt
@@ -1,38 +1,32 @@
 package com.abloom.mery.data.firebase.qna
 
-import com.abloom.mery.data.firebase.toTimestamp
-import com.google.firebase.Timestamp
-import com.google.firebase.firestore.PropertyName
+import androidx.annotation.Keep
+import com.abloom.mery.data.firebase.Document
+import com.abloom.mery.data.firebase.toTimestamp1
+import dev.gitlive.firebase.firestore.Timestamp
+import kotlinx.serialization.Serializable
 import java.time.LocalDateTime
 
-data class QnaDocument(
-    @JvmField @PropertyName(KEY_USER_ID) val userId: String = "",
-    @JvmField @PropertyName(KEY_QUESTION_ID) val questionId: Long = -1,
-    @JvmField @PropertyName(KEY_DATE) val date: Timestamp = Timestamp(0, 0),
-    @JvmField @PropertyName(KEY_ANSWER_CONTENT) val answer: String = "",
-    @JvmField @PropertyName(KEY_REACTION) val reaction: Int? = null,
-    @JvmField @PropertyName(KEY_IS_COMPLETE) val isComplete: Boolean = false,
-) {
+@Keep
+@Serializable
+class QnaDocument(
+    val user_id: String = "",
+    val q_id: Long = -1,
+    val date: Timestamp = Timestamp.now(),
+    val answer_content: String = "",
+    val reaction: Int? = null,
+    val is_complete: Boolean = false,
+) : Document {
 
-    companion object {
-
-        const val KEY_USER_ID = "user_id"
-        const val KEY_QUESTION_ID = "q_id"
-        const val KEY_DATE = "date"
-        const val KEY_ANSWER_CONTENT = "answer_content"
-        const val KEY_REACTION = "reaction"
-        const val KEY_IS_COMPLETE = "is_complete"
-
-        fun create(
-            userId: String,
-            questionId: Long,
-            date: LocalDateTime,
-            answer: String,
-        ) = QnaDocument(
-            userId = userId,
-            questionId = questionId,
-            date = date.toTimestamp(),
-            answer = answer,
-        )
-    }
+    constructor(
+        userId: String,
+        questionId: Long,
+        date: LocalDateTime,
+        answer: String,
+    ) : this(
+        user_id = userId,
+        q_id = questionId,
+        date = date.toTimestamp1(),
+        answer_content = answer,
+    )
 }
